electron/backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
numpy
opencv-python
insightface
onnxruntime
python-multipart

